+++ {{ cookiecutter.project_slug }}/apps/blog/__init__.py
@@ -1,1 +1,6 @@
-from .config import BlogConfig
+from django.apps import AppConfig
+
+
+class BlogConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'blog'
