+++ {{ cookiecutter.project_slug }}/apps/docs/AGENTS.md
