+++ {{ cookiecutter.project_slug }}/apps/docs/__init__.py
