+++ {{ cookiecutter.project_slug }}/apps/docs/admin.py
