+++ {{ cookiecutter.project_slug }}/apps/docs/content/deployment/docker-compose-deployment.md
