+++ {{ cookiecutter.project_slug }}/apps/docs/content/deployment/environment-variables.md
