+++ {{ cookiecutter.project_slug }}/apps/docs/content/features/example_feature.md
