+++ {{ cookiecutter.project_slug }}/apps/docs/content/getting-started/introduction.md
