+++ {{ cookiecutter.project_slug }}/apps/docs/models.py
