+++ {{ cookiecutter.project_slug }}/apps/docs/navigation.yaml
