+++ {{ cookiecutter.project_slug }}/apps/docs/tests.py
