+++ {{ cookiecutter.project_slug }}/apps/docs/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from docs.views import docs_page_view
+from apps.docs.views import docs_page_view
 
 urlpatterns = [
     path("<str:category>/<str:page>/", docs_page_view, name="docs_page"),
